backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
pydantic==2.9.2
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
google-generativeai==0.5.4